move mining scheduling to App.tsx
