+++ applications/launchpad/gui-react/src/index.tsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
+import { PersistGate } from 'redux-persist/integration/react'
 
 import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
-import { store } from './store'
+import { store, persistor } from './store'
 
 const container = document.getElementById('root')
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -13,7 +14,9 @@ const root = createRoot(container!)
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <PersistGate loading={null} persistor={persistor}>
+        <App />
+      </PersistGate>
     </Provider>
   </React.StrictMode>,
 )
