+++ applications/launchpad/gui-react/src/utils/useScheduling.ts
